upadate to solidity 0.8.1, refactor, add test commands, generate zkeys in repo with zkey manager and rust circom compiler
