+++ packages/contracts/config/small.json
@@ -1,0 +1,11 @@
+{
+  "stateTreeDepth": 10,
+  "intStateTreeDepth": 1,
+  "msgTreeDepth": 2,
+  "voteOptionTreeDepth": 2,
+  "msgBatchDepth": 1,
+  "circomPath": "\"/Users/Q/.cargo/bin/circom\"",
+  "snarkjsPath": "\"../../node_modules/snarkjs/build/cli.cjs\"",
+  "processMessagesCircom": "\"../../node_modules/qaci-circuits/circom/processMessages.circom\"",
+  "tallyVotesCircom": "\"../../node_modules/qaci-circuits/circom/tallyVotes.circom\""
+}
