+++ packages/contracts/contracts/BaseERC20Token.sol
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0
 
-pragma solidity ^0.7.2;
+pragma solidity ^0.8.1;
 
 import '@openzeppelin/contracts/token/ERC20/ERC20.sol';
