+++ packages/contracts/contracts/GrantRound.sol
@@ -1,15 +1,15 @@
 //SPDX-License-Identifier: MIT
 pragma experimental ABIEncoderV2;
-pragma solidity ^0.7.2;
+pragma solidity ^0.8.1;
 
-import {PollFactory, Poll, PollProcessorAndTallyer} from "maci-contracts/contracts/Poll.sol";
-import {VkRegistry} from "maci-contracts/contracts/VkRegistry.sol";
-import {Params} from "maci-contracts/contracts/Params.sol";
-import {IMACI} from "maci-contracts/contracts/IMACI.sol";
-import {AccQueue} from "maci-contracts/contracts/trees/AccQueue.sol";
+import {PollFactory, Poll, PollProcessorAndTallyer} from "qaci-contracts/contracts/Poll.sol";
+import {VkRegistry} from "qaci-contracts/contracts/VkRegistry.sol";
+import {Params} from "qaci-contracts/contracts/Params.sol";
+import {IMACI} from "qaci-contracts/contracts/IMACI.sol";
+import {AccQueue} from "qaci-contracts/contracts/trees/AccQueue.sol";
 
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
-import "@openzeppelin/contracts/token/ERC20/SafeERC20.sol";
+import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 
 import {IRecipientRegistry} from "./recipientRegistry/IRecipientRegistry.sol";
