+++ packages/contracts/contracts/QFI.sol
@@ -1,19 +1,19 @@
 // SPDX-License-Identifier: MIT
 pragma experimental ABIEncoderV2;
-pragma solidity ^0.7.2;
+pragma solidity ^0.8.1;
 
-import {MACI} from "maci-contracts/contracts/MACI.sol";
-import {Params} from "maci-contracts/contracts/Params.sol";
+import {MACI} from "qaci-contracts/contracts/MACI.sol";
+import {Params} from "qaci-contracts/contracts/Params.sol";
 
-import {PollFactory, Poll, PollProcessorAndTallyer, MessageAqFactory} from "maci-contracts/contracts/Poll.sol";
-import {VkRegistry} from "maci-contracts/contracts/VkRegistry.sol";
-import {InitialVoiceCreditProxy} from "maci-contracts/contracts/initialVoiceCreditProxy/InitialVoiceCreditProxy.sol";
-import {SignUpGatekeeper} from "maci-contracts/contracts/gatekeepers/SignUpGatekeeper.sol";
-import {ConstantInitialVoiceCreditProxy} from "maci-contracts/contracts/initialVoiceCreditProxy/ConstantInitialVoiceCreditProxy.sol";
-import {FreeForAllGatekeeper} from "maci-contracts/contracts/gatekeepers/FreeForAllSignUpGatekeeper.sol";
+import {PollFactory, Poll, PollProcessorAndTallyer, MessageAqFactory} from "qaci-contracts/contracts/Poll.sol";
+import {VkRegistry} from "qaci-contracts/contracts/VkRegistry.sol";
+import {InitialVoiceCreditProxy} from "qaci-contracts/contracts/initialVoiceCreditProxy/InitialVoiceCreditProxy.sol";
+import {SignUpGatekeeper} from "qaci-contracts/contracts/gatekeepers/SignUpGatekeeper.sol";
+import {ConstantInitialVoiceCreditProxy} from "qaci-contracts/contracts/initialVoiceCreditProxy/ConstantInitialVoiceCreditProxy.sol";
+import {FreeForAllGatekeeper} from "qaci-contracts/contracts/gatekeepers/FreeForAllSignUpGatekeeper.sol";
 
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
-import "@openzeppelin/contracts/token/ERC20/SafeERC20.sol";
+import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 
 import {GrantRound} from "./GrantRound.sol";
 import {FundsManager} from "./FundsManager.sol";
@@ -329,8 +329,8 @@ contract QFI is MACI, FundsManager {
 
         // The message batch size and the tally batch size
         BatchSizes memory batchSizes = BatchSizes(
-            MESSAGE_TREE_ARITY**uint8(_treeDepths.messageTreeSubDepth),
-            STATE_TREE_ARITY**uint8(_treeDepths.intStateTreeDepth)
+            uint8(MESSAGE_TREE_ARITY)**uint8(_treeDepths.messageTreeSubDepth),
+            uint8(STATE_TREE_ARITY)**uint8(_treeDepths.intStateTreeDepth)
         );
 
         GrantRound g = grantRoundFactory.deployGrantRound(
