+++ packages/contracts/contracts/recipientRegistry/OptimisticRecipientRegistry.sol
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0
 
-pragma solidity ^0.7.2;
+pragma solidity ^0.8.1;
 
 import '@openzeppelin/contracts/access/Ownable.sol';
 
@@ -103,7 +103,7 @@ contract OptimisticRecipientRegistry is Ownable, BaseRecipientRegistry {
     require(msg.value == baseDeposit, 'RecipientRegistry: Incorrect deposit amount');
     requests[recipientId] = Request(
       RequestType.Registration,
-      msg.sender,
+      payable(msg.sender),
       block.timestamp,
       msg.value,
       _recipient,
@@ -132,7 +132,7 @@ contract OptimisticRecipientRegistry is Ownable, BaseRecipientRegistry {
     require(msg.value == baseDeposit, 'RecipientRegistry: Incorrect deposit amount');
     requests[_recipientId] = Request(
       RequestType.Removal,
-      msg.sender,
+      payable(msg.sender),
       block.timestamp,
       msg.value,
       address(0),
