+++ packages/contracts/tests/QV/02-poll.ts
@@ -2,8 +2,8 @@ import { ethers } from "hardhat";
 import { BigNumber, BigNumberish, Signer } from "ethers";
 import chai from "chai";
 import { solidity } from "ethereum-waffle";
-import { Keypair, VerifyingKey } from "maci-domainobjs";
-import { G1Point, G2Point } from "maci-crypto";
+import { Keypair, VerifyingKey } from "qaci-domainobjs";
+import { G1Point, G2Point } from "qaci-crypto";
 
 import { PoseidonT3 } from "../../typechain/PoseidonT3";
 import { PoseidonT3__factory } from "../../typechain/factories/PoseidonT3__factory";
@@ -216,10 +216,10 @@ describe("Poll - New QV Voting Round", () => {
     PoseidonT6 = await PoseidonT6Factory.deploy();
 
     linkedLibraryAddresses = {
-      ["maci-contracts/contracts/crypto/Hasher.sol:PoseidonT5"]: PoseidonT5.address,
-      ["maci-contracts/contracts/crypto/Hasher.sol:PoseidonT3"]: PoseidonT3.address,
-      ["maci-contracts/contracts/crypto/Hasher.sol:PoseidonT6"]: PoseidonT6.address,
-      ["maci-contracts/contracts/crypto/Hasher.sol:PoseidonT4"]: PoseidonT4.address,
+      ["qaci-contracts/contracts/crypto/Hasher.sol:PoseidonT5"]: PoseidonT5.address,
+      ["qaci-contracts/contracts/crypto/Hasher.sol:PoseidonT3"]: PoseidonT3.address,
+      ["qaci-contracts/contracts/crypto/Hasher.sol:PoseidonT6"]: PoseidonT6.address,
+      ["qaci-contracts/contracts/crypto/Hasher.sol:PoseidonT4"]: PoseidonT4.address,
     };
 
     GrantRoundFactory = new GrantRoundFactory__factory({ ...linkedLibraryAddresses }, deployer);
@@ -408,7 +408,7 @@ describe("Poll - New QV Voting Round", () => {
 
     const expectedPollId = "0";
     //const expectedPollAddress = "0x4aAa546748B9b4DBaa3A1C58f0DD1b2d3C6E97Ef"; // NOTE: nonce hack to get the poll address. Use this for Integration test
-    const expectedPollAddress = "0xB550ae8163A275BD1a43D2074716FC384c696265"; // NOTE: nonce hack to get the poll address. Use this for Unit test
+    const expectedPollAddress = "0xb22F17625EC8fD0173098Ae0C04485FD3F11BD19"; // NOTE: nonce hack to get the poll address. Use this for Unit test
 
     const expectedCoordinatorPublicKey = Object.values(_coordinatorPubkey);
