+++ packages/contracts/tests/Unit/GrantRound.ts
@@ -13,16 +13,16 @@ import {
   PoseidonT4__factory,
   PoseidonT5__factory,
   PoseidonT6__factory,
-} from "../typechain";
-import BaseERC20TokenAbi from "../abi/@openzeppelin/contracts/token/ERC20/ERC20.sol/ERC20.json";
-import VkRegistryAbi from "../abi/maci-contracts/contracts/VkRegistry.sol/VkRegistry.json";
-import RecipientRegistryAbi from "../abi/contracts/recipientRegistry/OptimisticRecipientRegistry.sol/OptimisticRecipientRegistry.json";
-import QfiAbi from "../abi/contracts/QFI.sol/QFI.json";
-import MessageAqAbi from "../abi/maci-contracts/contracts/trees/AccQueue.sol/AccQueue.json";
-import AccQueueQuinaryBlankSlAbi from "../abi/maci-contracts/contracts/trees/AccQueue.sol/AccQueueQuinaryBlankSl.json";
-import { Command, Keypair } from "maci-domainobjs";
-import { MessageStruct } from "../typechain/GrantRound";
-import { hash4, hash5 } from "maci-crypto";
+} from "../../typechain";
+import BaseERC20TokenAbi from "../../abi/@openzeppelin/contracts/token/ERC20/ERC20.sol/ERC20.json";
+import VkRegistryAbi from "../../abi/qaci-contracts/contracts/VkRegistry.sol/VkRegistry.json";
+import RecipientRegistryAbi from "../../abi/contracts/recipientRegistry/OptimisticRecipientRegistry.sol/OptimisticRecipientRegistry.json";
+import QfiAbi from "../../abi/contracts/QFI.sol/QFI.json";
+import MessageAqAbi from "../../abi/qaci-contracts/contracts/trees/AccQueue.sol/AccQueue.json";
+import AccQueueQuinaryBlankSlAbi from "../../abi/qaci-contracts/contracts/trees/AccQueue.sol/AccQueueQuinaryBlankSl.json";
+import { Command, Keypair } from "qaci-domainobjs";
+import { MessageStruct } from "../../typechain/GrantRound";
+import { hash4, hash5 } from "qaci-crypto";
 
 chai.use(solidity);
 const { expect } = chai;
@@ -130,13 +130,13 @@ describe("Grant Round", () => {
     // ISSUE -> https://github.com/EthWorks/Waffle/issues/429.
     const GrantRoundFactory = new GrantRound__factory(
       {
-        ["maci-contracts/contracts/crypto/Hasher.sol:PoseidonT5"]:
+        ["qaci-contracts/contracts/crypto/Hasher.sol:PoseidonT5"]:
           poseidonT5.address,
-        ["maci-contracts/contracts/crypto/Hasher.sol:PoseidonT3"]:
+        ["qaci-contracts/contracts/crypto/Hasher.sol:PoseidonT3"]:
           poseidonT3.address,
-        ["maci-contracts/contracts/crypto/Hasher.sol:PoseidonT6"]:
+        ["qaci-contracts/contracts/crypto/Hasher.sol:PoseidonT6"]:
           poseidonT6.address,
-        ["maci-contracts/contracts/crypto/Hasher.sol:PoseidonT4"]:
+        ["qaci-contracts/contracts/crypto/Hasher.sol:PoseidonT4"]:
           poseidonT4.address,
       },
       deployer
